backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
python-dotenv
email-validator
psycopg2-binary
PyPDF2
pytesseract
Pillow
pdf2image
google-genai
anthropic
fpdf2
pymupdf
pytest-asyncio
chromadb
sentence-transformers